[UPD] Moved language select to own component
